+++ packages/helic/test/Helic/Test/ConfigFileTest.hs
@@ -1,5 +1,3 @@
-{-# options_ghc -Wno-unused-imports #-}
-
 module Helic.Test.ConfigFileTest where
 
 import Path (Abs, File, Rel, absfile, relfile)
@@ -11,15 +9,18 @@ import Helic.Config.File (parseFileConfig)
 import Helic.Data.Config (Config (Config))
 import Helic.Data.NetConfig (NetConfig (NetConfig))
 import Helic.Data.TmuxConfig (TmuxConfig (TmuxConfig))
+import Helic.Data.X11Config (X11Config (X11Config))
 
 target :: Config
 target =
-  Config (Just "name") (Just tmux) (Just net) (Just 1000) (Just False)
+  Config (Just "name") (Just tmux) (Just net) (Just x) (Just 1000) (Just False)
   where
     tmux =
       TmuxConfig (Just True) (Just [absfile|/bin/tmux|])
     net =
       NetConfig (Just 10001) (Just 5) (Just ["remote:1000"])
+    x =
+      X11Config (Just ":1")
 
 test_readConfigFile :: UnitTest
 test_readConfigFile = do